+++ frontend/App.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import * as Font from 'expo-font';
 import { setCustomText } from 'react-native-global-props';
 import { Fonts } from './assets/fonts/Fonts';
-import Map from '@screens/Map';
 import { AppContainer } from './components/Navigation';
 
 export default class App extends React.Component {
@@ -35,22 +34,6 @@ export default class App extends React.Component {
       return null;
     }
     return (
-      <View style={styles.container}>
-        <Map/>
-      </View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    width: '100%',
-    // backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
       <AppContainer />
     );
   }
